Peter's comments part 3
Fix up Document API docs when editor is not assigned.
Refactor error dialog UI from FileCommandHandlers to EditorManager.
Handle edge case when creating an editor for a file that no longer exists.
